% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.fun.R, R/huber.M.R
\name{alpha.fun}
\alias{alpha.fun}
\alias{huber.M}
\title{Utility functions for robust fitting}
\usage{
alpha.fun(r, k, q)

huber.M(data, q = pchisq(3, 2))
}
\arguments{
\item{r, k, q}{Parameters. For \code{huber.M}, \code{q=pchisq(k+1,k)}
gives the maximum breakdown point.}

\item{data}{A numeric matrix, one observation per row.}
}
\description{
Functions used by \code{\link{sma}} when 'robust = TRUE'.
}
\seealso{
\code{\link{sma}}
}
\author{
Warton, D. I. \email{David.Warton@unsw.edu.au}, S. Taskinen
}
\keyword{internal}
