% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineAxis.R
\name{defineAxis}
\alias{defineAxis}
\title{Defines axis for SMA object}
\usage{
defineAxis(
  major.ticks,
  limits = NULL,
  minor.ticks = NULL,
  major.tick.labels = major.ticks,
  both.sides = FALSE
)
}
\arguments{
\item{major.ticks}{values for major ticks}

\item{limits}{limits for axes}

\item{minor.ticks}{values for minor ticks}

\item{major.tick.labels}{labels for major ticks}

\item{both.sides}{deprecated}
}
\description{
Defines axis for SMA object
}
\keyword{internal}
