% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{ggplot_xy}
\alias{ggplot_xy}
\title{Make ggplot with transformed x and y axes}
\usage{
ggplot_xy(obj, pdat, ...)
}
\arguments{
\item{obj}{object with class sma}

\item{pdat}{plotting data generated by make_plot_data}

\item{...}{arguments passed to ggplot and geom_line}
}
\description{
Make ggplot with transformed x and y axes
}
