% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{ggplot_y}
\alias{ggplot_y}
\title{Make ggplot with transformed y axes only}
\usage{
ggplot_y(obj, pdat, ...)
}
\arguments{
\item{obj}{object with class sma}

\item{pdat}{plotting data generated by make_plot_data}

\item{...}{arguments passed to ggplot and geom_line}
}
\description{
Make ggplot with transformed y axes only
}
