% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik_sma.R
\name{logLik.sma}
\alias{logLik.sma}
\title{Computes log liklihood of SMA object}
\usage{
\method{logLik}{sma}(object, ...)
}
\arguments{
\item{object}{sma object}

\item{...}{}
}
\description{
Computes log liklihood of SMA object
}
