% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr.b.com.R
\name{lr.b.com}
\alias{lr.b.com}
\title{Likelihood ratio statistic, assuming a common slope (called by slope.com)}
\usage{
lr.b.com(b, arguments)
}
\arguments{
\item{b}{Common slope estimate.}

\item{arguments}{A list of other required arguments.}
}
\description{
Likelihood ratio statistic, assuming a common slope (called by slope.com)
}
\seealso{
\code{\link{slope.com}}
}
\author{
Warton, D. \email{David.Warton@unsw.edu.au}, translated to R by
Ormerod, J. 2005-12-08
}
\keyword{internal}
