% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma.R
\name{ma}
\alias{ma}
\title{Major axis estimation and testing for one or several samples}
\usage{
ma(...)
}
\arguments{
\item{...}{arguments passed to sma()}
}
\description{
Major axis estimation and testing for one or several samples
}
