% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{make_data_raw}
\alias{make_data_raw}
\title{Reformat data for plotting on untransformed data}
\usage{
make_data_raw(obj)
}
\arguments{
\item{obj}{ma or sma obj}
}
\value{
tibble of plotting data
}
\description{
Reformat data for plotting on untransformed data
}
\keyword{internal}
