% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{make_data_x}
\alias{make_data_x}
\title{Reformat data for plotting on log transformed x}
\usage{
make_data_x(obj)
}
\arguments{
\item{obj}{ma or sma obj}
}
\description{
Reformat data for plotting on log transformed x
}
