% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{make_data_xy}
\alias{make_data_xy}
\title{Reformat data for plotting on log transformed x and y}
\usage{
make_data_xy(obj)
}
\arguments{
\item{obj}{ma or sma obj}
}
\description{
Reformat data for plotting on log transformed x and y
}
