% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{make_data_y}
\alias{make_data_y}
\title{Reformat data for plotting on log transformed y}
\usage{
make_data_y(obj)
}
\arguments{
\item{obj}{ma or sma obj}
}
\description{
Reformat data for plotting on log transformed y
}
