% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLogMinor.R
\name{makeLogMinor}
\alias{makeLogMinor}
\title{Generate spacing for minor tick marks on a plot with log-scaled axes}
\usage{
makeLogMinor(major)
}
\arguments{
\item{major}{a vector of values giving major tick marks.}
}
\value{
vector of minor tick spacings appropriate for log 10 scaled axis with major ticks given by 'major'
}
\description{
Generates a sequence of numbers providing minor tick marks on log scaled
axes.

A vector is created according to the following algorithm. For each pair of
adjacent values (x1, x2) in \code{major}, the function adds the values (x1,
2*x1, 3*x1, ..., x2) to the vector of return values.

This is useful for generating spacing of minor tick-values on
log-transformed axes.
}
\examples{
#Sequence suitable for log base 10 labels
makeLogMinor(seqLog(1E-5, 1E5))
}
\seealso{
\code{\link{plotutils}}
}
