% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nicePlot.R
\name{nicePlot}
\alias{nicePlot}
\title{Creates nice plot}
\usage{
nicePlot(
  xaxis,
  yaxis,
  log = "",
  ann = par("ann"),
  xlab = NULL,
  ylab = NULL,
  tck = par("tck"),
  frame.plot = TRUE,
  ...
)
}
\arguments{
\item{xaxis}{}

\item{yaxis}{}

\item{log}{}

\item{ann}{}

\item{xlab}{}

\item{ylab}{}

\item{tck}{}

\item{frame.plot}{}

\item{...}{}
}
\description{
Creates nice plot
}
