% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.sma.R
\name{predict.sma}
\alias{predict.sma}
\title{Predicting from an (Standardized) major axis estimation}
\usage{
\method{predict}{sma}(object, ...)
}
\arguments{
\item{object}{sma obj}

\item{...}{}
}
\value{
A cautionary message
}
\description{
Predicting from an (Standardized) major axis estimation
}
