\name{print.sma}
\alias{print.sma}
\title{Print an object of class 'sma'.}
\description{Prints the coefficients and the results of the various hypothesis tests of an sma object.}
\usage{
\method{print}{sma}(x, ..., coefbygroup = FALSE)
}
\arguments{
  \item{x}{Object of class 'sma'.}
  \item{coefbygroup}{Whether or not to print the coefficients by group.}
  \item{\dots}{Further arguments ignored.}
  
}

\author{ R.A. Duursma, D. Falster,  D.I. Warton }
\seealso{
\code{\link{sma}}
}
\keyword{misc}
