% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqLog.R
\name{seqLog}
\alias{seqLog}
\title{Sequence Generation}
\usage{
seqLog(from, to, base = 10)
}
\arguments{
\item{from, to}{the starting and (maximal) end value of a sequence.}

\item{base}{multiplication value.}
}
\value{
vector from lo to hi with multiplication steps of incr. Used for making ticks to a log-scaled axis
}
\description{
Generate multiplicative sequences, or series.
}
\details{
Starting at \code{from}, \code{seq} multiplies successively by \code{base}
until the maximal value is reached. This is useful for generating
tick-spacing on log-transformed axes.
}
\examples{
#Sequence suitable for log base 10 labels
seqLog(1E-5, 1E5)

#Sequence suitable for log base 2 labels
seqLog(2, 128,base=2)
}
\seealso{
\code{\link{plotutils}}
}
