% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.sma.R
\name{summary.sma}
\alias{summary.sma}
\title{Print a summary}
\usage{
\method{summary}{sma}(object, ...)
}
\arguments{
\item{object}{Object of class 'sma'.}

\item{\dots}{Further arguments ignored.}
}
\description{
Writes a summary of a (standardized) major axis fit.

See \code{\link{sma}} for examples.
}
\seealso{
\code{\link{sma}},\code{\link{print.sma}}
}
\author{
Remko Duursma, Daniel Falster, David Warton
}
\keyword{misc}
