% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.fun.R
\name{alpha.fun}
\alias{alpha.fun}
\title{Utility functions for robust fitting}
\usage{
alpha.fun(r, k, q)
}
\arguments{
\item{r, k, q, }{Parameters.}

\item{data}{.....}
}
\description{
Functions used by \code{\link{sma}} when 'robust = TRUE'.
}
\seealso{
\code{\link{sma}}
}
\author{
Warton, D. I. \email{David.Warton@unsw.edu.au}, S. Taskinen
}
\keyword{internal}
