% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.sma.R
\name{coef.sma}
\alias{coef.sma}
\title{Extract coefficients from a 'sma' or 'ma' fit}
\usage{
\method{coef}{sma}(object, ...)
}
\arguments{
\item{object}{Object of class 'sma'.}

\item{\dots}{Further arguments ignored.}
}
\value{
A dataframe with the slope(s) and elevation(s), and their confidence
intervals. If the fit was by multiple groups, fits by all groups are
returned.
}
\description{
Extracts elevation and slope of a standardized major axis (sma) or major
axis (ma) fit, for each of the groups if the fit was by group.
}
\seealso{
\code{\link{sma}}
}
\author{
R.A. Duursma
}
\keyword{misc}
