% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.sma.R
\name{confint.sma}
\alias{confint.sma}
\title{Returns confidence intervals for sma object}
\usage{
\method{confint}{sma}(object, ...)
}
\arguments{
\item{object}{a sma object}

\item{...}{arguments passed}
}
\value{
dataframe containing confidence interval
}
\description{
Returns confidence intervals for sma object
}
