% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineAxis.R
\name{defineAxis}
\alias{defineAxis}
\title{Defines axis for SMA object}
\usage{
defineAxis(
  major.ticks,
  limits = NULL,
  minor.ticks = NULL,
  major.tick.labels = major.ticks,
  both.sides = FALSE
)
}
\arguments{
\item{major.ticks}{}

\item{limits}{}

\item{minor.ticks}{}

\item{major.tick.labels}{}

\item{both.sides}{}
}
\description{
Defines axis for SMA object
}
\keyword{internal}
