% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust.factor.R
\name{gamma.fun}
\alias{gamma.fun}
\title{Gamma function used in robust factor calculations}
\usage{
\method{gamma}{fun}(r, k, q)
}
\arguments{
\item{r}{}

\item{k}{}

\item{q}{}

\item{data}{}
}
\value{
Computers robust factor for elevation, slope and CI
}
\description{
Gamma function used in robust factor calculations
}
\keyword{internal}
