% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{get_coef}
\alias{get_coef}
\title{Extract coefficients for each level of group}
\usage{
get_coef(obj, group_level, coef_type)
}
\arguments{
\item{obj}{object with sma class}

\item{group_level}{character string of group level}

\item{coef_type}{type of coefficient you want a for intercept B for slope}
}
\description{
Extract coefficients for each level of group
}
