% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{ggplot.sma}
\alias{ggplot.sma}
\title{Plot (S)MA using ggplot2}
\usage{
\method{ggplot}{sma}(obj, ...)
}
\arguments{
\item{obj}{object with class sma}

\item{...}{arguments passed to geom_line()}
}
\value{
ggplot2 obj
}
\description{
Plot (S)MA using ggplot2
}
