% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huber.M.R
\name{huber.M}
\alias{huber.M}
\title{Huber's M-estimator}
\usage{
huber.M(data, q = pchisq(3, 2))
}
\arguments{
\item{data}{}

\item{q}{q=pchisq(k+1,k) gives maximum breakdown point}
}
\description{
Huber's M-estimator
}
\keyword{internal}
