\name{lr.b.com}
\alias{lr.b.com}
\title{ Likelihood ratio statistic, assuming a common slope (called by slope.com) }
\description{
  Likelihood ratio statistic, assuming a common slope (called by slope.com)
}
\usage{
lr.b.com(b, arguments)
}
\arguments{
  \item{b}{ Common slope estimate. }
  \item{arguments}{ A list of other required arguments. }
}
\author{ Warton, D. \email{David.Warton@unsw.edu.au}, translated to R by Ormerod, J. 2005-12-08 }

\seealso{ \code{\link{slope.com}} }
\keyword{ internal }
