% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.sma.R
\name{print.sma}
\alias{print.sma}
\title{Print an object of class 'sma'.}
\usage{
\method{print}{sma}(x, ..., coefbygroup = FALSE)
}
\arguments{
\item{x}{Object of class 'sma'.}

\item{\dots}{Further arguments ignored.}

\item{coefbygroup}{Whether or not to print the coefficients by group.}
}
\description{
Prints the coefficients and the results of the various hypothesis tests of
an sma object.
}
\seealso{
\code{\link{sma}}
}
\author{
R.A. Duursma, D. Falster, D.I. Warton
}
\keyword{misc}
