% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust.factor.R
\name{robust.factor}
\alias{robust.factor}
\title{Computers robust factor for elevation, slope and CI}
\usage{
robust.factor(data, q)
}
\arguments{
\item{data}{}

\item{q}{}
}
\description{
Computers robust factor for elevation, slope and CI
}
\keyword{internal}
