\name{summary.sma}
\alias{summary.sma}
\title{Print a summary}
\description{Writes a summary of a (standardized) major axis fit.
}
\usage{
\method{summary}{sma}(object, ...)
}
\arguments{
  \item{object}{Object of class 'sma'.}
  \item{\dots}{Further arguments ignored.}
}
\details{
Does not add much to print.sma, at the moment, except when fit by multiple groups.
See \code{\link{sma}} for examples.
}
\author{Remko Duursma, Daniel Falster, David Warton
}

\seealso{
\code{\link{sma}},\code{\link{print.sma}}
}

\keyword{misc}
